requirements.txt
fastapi
uvicorn
requests
requests_oauthlib
python-dotenv
